:fire: Remove unwanted duplicate test file and rename actual tests
This came about in an upstream commit that probably shouldn't have existed.
