+++ test/test_bin/test_identify_dormant_github_users.py
@@ -1,71 +1,123 @@
-from unittest import TestCase
-from unittest.mock import patch, MagicMock, Mock
+import unittest
+from datetime import datetime, timedelta
+from unittest.mock import MagicMock, mock_open, patch
 
-from bin.identify_dormant_github_users import (get_audit_logs, identify_dormant_users, identify_dormant_github_users, setup_environment)
-from services.github_service import GithubService
+from botocore.exceptions import ClientError, NoCredentialsError
+
+from bin.identify_dormant_github_users import (
+    calculate_date_by_integer, dormant_users_according_to_github,
+    dormant_users_not_in_auth0_audit_log,
+    download_github_dormant_users_csv_from_s3,
+    get_usernames_from_csv_ignoring_bots, setup_services)
 from services.auth0_service import Auth0Service
-from config.constants import RESPONSE_OKAY
-
-import requests
+from services.github_service import GithubService
 
 
-class TestIdentifyDormantGithubUsers(TestCase):
+class TestDormantGitHubUsers(unittest.TestCase):
 
-    def setUp(self) -> None:
-        self.response = Mock()
-        self.response.status_code = RESPONSE_OKAY
-        self.response.json.return_value = {"access_token": "test_access_token"}
-        requests.post = MagicMock(return_value=self.response)
+    def setUp(self):
+        self.allowed_bot_users = ["bot1", "bot2"]
 
-    @patch("sys.argv", ["test_github_token", "test_client_secret", "test_client_id", "test_domain"])
-    def test_setup_environment(self):
-        self.assertEqual(setup_environment(), ("test_github_token", "test_client_secret", "test_client_id", "test_domain"))
+    def test_calculate_date_by_integer_valid_days(self):
+        result = calculate_date_by_integer(7)
+        expected_date = datetime.now() - timedelta(days=7)
+        expected_date = expected_date.replace(
+            hour=0, minute=0, second=0, microsecond=0)
+        self.assertEqual(result, expected_date)
 
-    @patch.object(Auth0Service, "get_active_users")
-    def test_get_audit_logs(self, mock_auth0_users):
-        mock_auth0_users.return_value = [
-            {
-                "test_key": "test_value",
-                "last_login": "2050-01-01T00:00:00.000Z",
-                "user_id": "user1"
-            },
-            {
-                "test_key": "test_value",
-                "last_login": "2050-01-01T00:00:00.000Z",
-                "user_id": "user2"
-            }
-        ]
+    def test_calculate_date_by_integer_zero_days(self):
+        result = calculate_date_by_integer(0)
+        expected_date = datetime.now()
+        expected_date = expected_date.replace(
+            hour=0, minute=0, second=0, microsecond=0)
+        self.assertEqual(result, expected_date)
 
-        auth0_service = Auth0Service("test_client_secret", "test_client_id", "test_domain", "test_grant_type")
+    def test_calculate_date_by_integer_negative_days(self):
+        result = calculate_date_by_integer(-7)
+        expected_date = datetime.now() + timedelta(days=7)
+        expected_date = expected_date.replace(
+            hour=0, minute=0, second=0, microsecond=0)
+        self.assertEqual(result, expected_date)
 
-        self.assertEqual(get_audit_logs(auth0_service), ["user1", "user2"])
+    def test_get_usernames_from_csv_ignoring_bots_empty_csv(self):
+        mock_file_content = ""
+        with patch("builtins.open", mock_open(read_data=mock_file_content)):
+            result = get_usernames_from_csv_ignoring_bots(
+                self.allowed_bot_users)
+        self.assertEqual(result, [])
 
-    def test_identify_dormant_users(self):
-        self.assertEqual(identify_dormant_users(["user1", "user2", "user3"], ["user1", "user2"]), ["user3"])
+    def test_get_usernames_from_csv_ignoring_bots_only_bots(self):
+        mock_file_content = "ci-hmcts\ncloud-platform-dummy-user\n"
+        with patch("builtins.open", mock_open(read_data=mock_file_content)):
+            result = get_usernames_from_csv_ignoring_bots(
+                self.allowed_bot_users)
+        self.assertEqual(result, [])
 
-    @patch("gql.transport.aiohttp.AIOHTTPTransport.__new__", new=MagicMock)
-    @patch("gql.Client.__new__", new=MagicMock)
-    @patch("github.Github.__new__")
-    @patch("bin.identify_dormant_github_users.setup_environment")
-    @patch("bin.identify_dormant_github_users.get_audit_logs")
-    @patch.object(GithubService, "get_users_of_multiple_organisations")
-    @patch("bin.identify_dormant_github_users.identify_dormant_users")
-    def test_identify_dormant_github_users(
-        self,
-        mock_identify_dormant_users,
-        mock_get_users_if_multiple_organisations,
-        mock_get_audit_logs,
-        mock_setup_environment,
-        _mock_github_client_core_api
-    ):
-        mock_setup_environment.return_value = ("test_github_token", "test_client_secret", "test_client_id", "test_domain")
-        mock_get_audit_logs.return_value = ["user1", "user2", "user3"]
-        mock_get_users_if_multiple_organisations.return_value = ["user1", "user2", "user3", "user4"]
-        mock_identify_dormant_users.return_value = ["user4"]
+    @patch('boto3.client')
+    def test_download_github_dormant_users_csv_from_s3_no_credentials(self, mock_boto3_client):
+        # Mock boto3 client to raise NoCredentialsError
+        mock_boto3_client.side_effect = NoCredentialsError
+        with self.assertRaises(NoCredentialsError):
+            download_github_dormant_users_csv_from_s3()
 
-        identify_dormant_github_users()
+    @patch('boto3.client')
+    def test_download_github_dormant_users_csv_from_s3_file_not_found(self, mock_boto3_client):
+        # Mock boto3 client to raise ClientError for a not found file
+        error_response = {'Error': {'Code': '404', 'Message': 'Not Found'}}
+        mock_boto3_client.side_effect = ClientError(
+            error_response, 'get_object')
+        with self.assertRaises(ClientError):
+            download_github_dormant_users_csv_from_s3()
 
-        mock_setup_environment.assert_called_once()
-        mock_get_audit_logs.assert_called_once()
-        mock_get_users_if_multiple_organisations.assert_called_once_with(["ministryofjustice", "moj-analytical-services"])
-        mock_identify_dormant_users.assert_called_once_with(["user1", "user2", "user3", "user4"], ["user1", "user2", "user3"])
+    @patch.object(GithubService, "__new__")
+    @patch('bin.identify_dormant_github_users.get_dormant_users_from_github_csv')
+    def test_dormant_users_according_to_github_all_active(self, mock_get_dormant_users_from_github_csv, mock_github_service):
+        # Mock `get_dormant_users_from_github_csv` to return a list of users
+        mock_get_dormant_users_from_github_csv.return_value = [
+            'user1', 'user2']
+        # Mock `github_service` to return an empty set
+        mock_github_service.check_dormant_users_audit_activity_since_date.return_value = set()
+        result = dormant_users_according_to_github(
+            mock_github_service, datetime.now() - timedelta(days=30))
+        self.assertEqual(result, set())
+
+    @patch.object(GithubService, "__new__")
+    @patch('bin.identify_dormant_github_users.get_dormant_users_from_github_csv')
+    def test_dormant_users_according_to_github_all_dormant(self, mock_get_dormant_users_from_github_csv, mock_github_service):
+        # Mock `get_dormant_users_from_github_csv` to return a list of users
+        mock_get_dormant_users_from_github_csv.return_value = [
+            'user1', 'user2']
+        # Mock `github_service` to return a set of users
+        mock_github_service.check_dormant_users_audit_activity_since_date.return_value = {
+            'user1', 'user2'}
+        result = dormant_users_according_to_github(
+            mock_github_service, datetime.now() - timedelta(days=30))
+        self.assertEqual(result, {'user1', 'user2'})
+
+    def test_dormant_users_not_in_auth0_audit_log_all_present(self):
+        # Mock `auth0_service` to return a set of users
+        mock_auth0_service = MagicMock(spec=Auth0Service)
+        mock_auth0_service.get_active_case_sensitive_usernames.return_value = {
+            'user1', 'user2'}
+        result = dormant_users_not_in_auth0_audit_log(
+            mock_auth0_service, {'user1', 'user2'})
+        self.assertEqual(result, set())
+
+    def test_dormant_users_not_in_auth0_audit_log_all_absent(self):
+        # Mock `auth0_service` to return an empty set
+        mock_auth0_service = MagicMock(spec=Auth0Service)
+        mock_auth0_service.get_active_case_sensitive_usernames.return_value = []
+        result = dormant_users_not_in_auth0_audit_log(
+            mock_auth0_service, {'user1', 'user2'})
+        self.assertEqual(result, {'user1', 'user2'})
+
+    @patch('os.environ.get')
+    def test_setup_services_missing_gh_admin_token(self, mock_env_get):
+        # Mock `os.environ.get` to return `None` for `GH_ADMIN_TOKEN`
+        mock_env_get.side_effect = lambda k: None if k == 'GH_ADMIN_TOKEN' else 'dummy_value'
+        with self.assertRaises(ValueError):
+            setup_services()
+
+
+if __name__ == '__main__':
+    unittest.main()
